feat(build): spyc-vt-sys — pinned libghostty-vt FFI with vendored archives (#459)
Stage 4 of the engine series. A new workspace crate owning everything C.
spyc
does not depend on it yet — the flip is PR 15.

## The pin

`1f5bb5769fbb5e717546073d33d3985604a315b2`, 2026-09-04 — a **bare
commit, with
its date and rationale recorded beside the hash**, because no tag
qualifies:

- No released ghostty tag carries the C API at all. `v1.3.1`, the
latest, has no
`include/ghostty/vt/` and requires Zig 0.15.2, which cannot link on
macOS 26.
- Ghostty publishes a rolling `tip` tag. It had already moved off this
commit by
the time I recorded the pin — a tag that moves is strictly worse than a
hash.

It is the commit every Stage 4/5 figure is measured at.

## The bindings are generated, not borrowed

The published `libghostty-vt-sys 0.2.1` bindings are
**ABI-incompatible** with
this pin: `ghostty_terminal_new` lost its options struct in favour of
two scalar
parameters, and those bindings still compile against the newer library
and
return garbage — `rows()` reports the scrollback budget. A C ABI has no
version
handshake.

So `src/bindings.rs` is bindgen output from **this pin's** headers,
checked in so
neither bindgen nor libclang is a crate dependency. It carries **~371
compile-time layout assertions** — a struct that changes shape fails the
build
instead of lying. `use_core()` is deliberately not set: it would
silently drop
those assertions, which name `::std::mem`.

## The byte ceiling, as decided

Rows are the UX contract; bytes are the safety valve. Neither is left at
its
default — the default byte cap truncates history to ~840 rows
*irrespective of
the line limit*, so a 10,000-row budget would silently deliver 8% of
itself.

The derivation is code (`src/scrollback.rs`), not prose:
`rows × BYTES_PER_ROW_HEAVY + 2 pages`.

- **`BYTES_PER_ROW_HEAVY = 929`**, measured at the pin on an
agent-frame-shaped
stream — several SGR runs per row including truecolour, box drawing, a
wide
  char, an emoji, a combining mark — not on build-log spew.
- Worth recording: heavy and light content differ by only ~6% (929 vs
767 B/row),
because ghostty's page storage is dominated by fixed per-row cost rather
than
style content. The two corpora agreeing *exactly* at the tightest
ceiling is
page quantisation showing through, which is why the derivation takes the
  **max** rather than the mean.
- **Slack is additive, not multiplicative** — page granularity is a
fixed
quantum that does not scale with the budget, so a percentage margin
would
  over-provision a large budget and under-provision a small one.
- At the default 10k budget: **10,109,200 B = 9.64 MiB/pane**, pinned by
a test
that asserts the exact value so a change has to be deliberate, and well
under
the ~50 MB threshold that would instead prompt revisiting the row
budget.
- A separate pathological measurement (7,828 B/row) exists to prove the
valve
  *binds*; it is not an input to the derivation. That becomes Stage 5's
  second gate assertion.

## Five archives, not four

The release matrix is macOS arm64/x86_64 and musl aarch64/x86_64 — but
**CI's
Lint and Tests jobs run on `ubuntu-latest`, whose host target is
`x86_64-unknown-linux-gnu` and is not a release artifact.** A
four-archive set
fails the gate on the PR that adds it. `build.rs`'s error message says
so.

## Checksums, bitten both ways

Every archive's SHA-256 is verified before a single link flag is
emitted, with a
local SHA-256 so a build script that verifies a checksum does not itself
pull a
dependency tree. Verified rather than assumed:

- one flipped byte → build fails with both hashes printed;
- a missing `CHECKSUMS` entry → fails loudly rather than skipping
verification;
- restore → builds clean.

## Two crates publish, in order

`spyc-vt-sys` must reach crates.io **before** spyc. `cargo package`
refuses a
dependency without a registry version — verified by adding the Stage-15
dependency and watching it fail. They package separately, which is what
makes
the archives affordable: `spyc-vt-sys` measures **3.91 MiB** as a real
`.crate`
and spyc's own stays ~1.6 MiB, each independently inside the 10 MiB cap
rather
than sharing one budget. Both from an actual `cargo package`, not
arithmetic.
Recorded in `docs/RELEASE_ENGINEERING.md`.

## The release job verifies rather than builds

`vendored-archives` **rebuilds** the archives from the pinned source and
diffs
the checksums. The archives are committed so `cargo install spyc` needs
no Zig;
the only thing standing behind SECURITY.md's "reproducible by anyone"
claim is
something actually reproducing them. `publish` now depends on it.

It pins a **different Zig from the `linux` job** — that one is
`cargo-zigbuild`'s cross-linker (0.14.0), this one must match ghostty's
`requireZig`, which compares major.minor for *equality*. The two uses
are
independent; the job comment says so.

## Upstream errands

Both ghostty emit bugs were re-checked **at the shipping pin before
filing**, as
instructed:

- **`with_tabstops(true)` clobbering the restored cursor: FIXED
upstream.** The
emit now places the cursor restore last; cursor round-trips
`(3,0)→(3,0)`.
  Nothing filed. This goes in the addendum instead.
- **The one-row viewport offset: reduced, not gone.** The cursor is now
correct,
but scrollback still goes 1977→1976. Whether that −1 is a real row loss
or a
counting boundary is what Stage 5's row-by-row re-grading settles, so I
would
rather file it then with a precise reproducer than now with a vague one.
- **wezterm's fuzz off-by-one: filed as

[wezterm/wezterm#8134](https://github.com/wezterm/wezterm/issues/8134)**,
delta-debugged from 734 bytes to **5** (`\x1b[;9H\x88`) with the
mechanism
isolated: CUP past the right edge leaves the cursor at `x == cols`, and
the
subsequent HTS indexes the tab-stop array there. Neither sequence panics
alone. The `--dump` / `--shrink` / `--variants` modes that produced that
are
the second commit here — a fuzz finding is only useful upstream if it
can be
  handed over without this crate attached.

## Notes on things I got wrong along the way

- I hand-rolled the formatter options struct to re-check the emit bugs
and it
**segfaulted** — I had put `selection` before `extra`. Redone against
the
generated bindings, which have the real order and the layout assertions
to
back it. Third time this session that hand-rolling something I could
have read
  bit me; the bindings exist for exactly this reason.
- The vendor recipe deliberately does **not** strip. An earlier version
did, and
on Linux the step was a silent no-op because `llvm-strip` was absent and
the
failure was swallowed by `|| true`. At `ReleaseSmall` the archives carry
no
debug sections and stripping changes nothing measurable, so the recipe
asserts
a per-target **size budget** instead — which is what actually keeps the
crate
  under the cap.

## Gate

`make check-ci`: `=== GATE: PASS ===`. `make lint-workflows` clean
(actionlint
1.7.12, shellcheck 0.11.0 — matching CI's pin). `make test` and `make
lint` now
pass `--workspace`; without it neither would touch the new crate at all.


---------

+++ spikes/vt-engine/src/engines/ghostty_engine.rs
@@ -22,7 +22,7 @@ pub struct GhosttyEngine {
 fn color(c: StyleColor) -> Color {
     match c {
         StyleColor::None => Color::Default,
-        StyleColor::Palette(i) => Color::Idx(i.0 as u8),
+        StyleColor::Palette(i) => Color::Idx(i.0),
         StyleColor::Rgb(rgb) => Color::Rgb(rgb.r, rgb.g, rgb.b),
     }
 }
